requirements.txt
streamlit==1.28.0
tensorflow==2.12.0
numpy==1.24.0
pandas==2.0.0
matplotlib==3.7.0
seaborn==0.12.0
pillow==10.0.0
opencv-python-headless==4.8.0
scipy==1.11.0
statsmodels==0.14.0
fpdf2==2.7.0
scikit-learn==1.0.0